refactor: keep fusion helper in search module
